Move global var and function to globalvar.go file.
Move apisql and apiapi function to dedicated package to make srvhttp lighter.

Define   "listening_port": 8081 in config-sample.json
