+++ internal/srvhttp/srvhttp.go
@@ -1,298 +1,15 @@
 package srvhttp
 
 import (
+	"apigw/internal/apiapi"
+	"apigw/internal/apisql"
+	"apigw/internal/globalvar"
 	"apigw/internal/logger"
-	"database/sql"
-	"encoding/json"
 	"fmt"
 	"github.com/fsnotify/fsnotify"
-	_ "github.com/go-sql-driver/mysql"
-	"github.com/gorilla/mux"
-	_ "github.com/lib/pq"
-	_ "github.com/mattn/go-sqlite3"
 	"github.com/spf13/viper"
-	"io"
 	"net/http"
-	"strings"
 )
-
-var sm = mux.NewRouter()
-var getR = sm.Methods(http.MethodGet).Subrouter()
-
-type sqlData struct {
-	ReturnedRows int64         `json:"ReturnedRows"`
-	Data         []interface{} `json:"Data"`
-}
-
-func httpClose(c io.Closer) {
-	err := c.Close()
-	checkErr(err)
-}
-
-func createApiSql(apiName string) {
-	logger.LogMsg(fmt.Sprintf("Requested api endpoint : %s", apiName), "info")
-
-	apiMethod := viper.GetString(fmt.Sprintf("api.%s.method", apiName))
-
-	if apiMethod == "get" {
-		getR.HandleFunc(apiName, querySql)
-		logger.LogMsg(fmt.Sprintf("Created GET api endpoint : %s", apiName), "info")
-	}
-
-}
-
-func querySql(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	apiName := r.URL
-	logger.LogMsg(fmt.Sprintf("Call %s", apiName), "info")
-
-	if viper.IsSet(fmt.Sprintf("api.%s", apiName)) {
-
-		dbQuery := viper.GetString(fmt.Sprintf("api.%s.query", apiName))
-		dbSource := viper.GetString(fmt.Sprintf("api.%s.source", apiName))
-		dbName := viper.GetString(fmt.Sprintf("sources.%s.dbname", dbSource))
-		dbUsername := viper.GetString(fmt.Sprintf("sources.%s.username", dbSource))
-		dbPassword := viper.GetString(fmt.Sprintf("sources.%s.password", dbSource))
-		dbHost := viper.GetString(fmt.Sprintf("sources.%s.host", dbSource))
-		dbPort := viper.GetInt(fmt.Sprintf("sources.%s.port", dbSource))
-		dbDriver := viper.GetString(fmt.Sprintf("sources.%s.engine", dbSource))
-
-		var db *sql.DB
-		var err error
-		var sqlData sqlData
-		var rowCount int64
-		var dbConnect bool
-		var endResponse *strings.Reader
-
-		switch strings.ToLower(dbDriver) {
-		case "sqlite":
-			db, err = sql.Open("sqlite3", dbName)
-			checkErr(err)
-			logger.LogMsg(fmt.Sprintf("Open sqlite db %s", dbName), "info")
-			dbConnect = true
-		case "mysql":
-			dbCnxString := fmt.Sprintf("%s:%s@tcp(%s:%d)/%s", dbUsername, dbPassword, dbHost, dbPort, dbName)
-			db, err = sql.Open("mysql", dbCnxString)
-			checkErr(err)
-
-			if err != nil {
-				logger.LogMsg(fmt.Sprintf("Can't open mysql db %s", dbName), "info")
-				dbConnect = false
-				endResponse = strings.NewReader(
-					fmt.Sprintf("{\"error\" : \"Sorry the call %s is unable to join the target endpoint, "+
-						"please contact an administrator\"}\n",
-						apiName,
-					),
-				)
-			} else {
-				logger.LogMsg(fmt.Sprintf("Open mysql db %s", dbName), "info")
-				dbConnect = true
-			}
-
-		case "postgres":
-			dbCnxString := fmt.Sprintf("postgresql://%s:%s@%s:%d/%s", dbUsername, dbPassword, dbHost, dbPort, dbName)
-			db, err = sql.Open("postgres", dbCnxString)
-			checkErr(err)
-
-			if err != nil {
-				logger.LogMsg(fmt.Sprintf("Can't open postgres db %s", dbName), "info")
-				dbConnect = false
-				endResponse = strings.NewReader(
-					fmt.Sprintf("{\"error\" : \"Sorry the call %s is unable to join the target endpoint, "+
-						"please contact an administrator\"}\n",
-						apiName,
-					),
-				)
-			} else {
-				logger.LogMsg(fmt.Sprintf("Open postgres db %s", dbName), "info")
-				dbConnect = true
-			}
-
-		default:
-			dbConnect = false
-			logger.LogMsg(
-				fmt.Sprintf(
-					"Sorry the call %s is misconfigured sql driver %s is not supported, ",
-					apiName, dbDriver,
-				),
-				"info",
-			)
-			endResponse = strings.NewReader(
-				fmt.Sprintf("{\"error\" : \"Sorry the call %s was misconfigured contact the administrator\"}\n",
-					apiName,
-				),
-			)
-
-		}
-
-		if dbConnect {
-			rows, err := db.Query(dbQuery)
-			checkErr(err)
-			logger.LogMsg(fmt.Sprintf("execute query : %s", dbQuery), "info")
-
-			columnTypes, err := rows.ColumnTypes()
-			checkErr(err)
-
-			colCount := len(columnTypes)
-			rowCount = 0
-
-			for rows.Next() {
-
-				scanArgs := make([]interface{}, colCount)
-
-				for i, v := range columnTypes {
-
-					switch v.DatabaseTypeName() {
-					case "VARCHAR", "TEXT", "UUID", "TIMESTAMP":
-						scanArgs[i] = new(sql.NullString)
-						break
-					case "BOOL":
-						scanArgs[i] = new(sql.NullBool)
-						break
-					case "INT4", "INTEGER", "numeric":
-						scanArgs[i] = new(sql.NullInt64)
-						break
-					default:
-						scanArgs[i] = new(sql.NullString)
-					}
-				}
-
-				err := rows.Scan(scanArgs...)
-				checkErr(err)
-
-				masterData := map[string]interface{}{}
-
-				for i, v := range columnTypes {
-
-					if z, ok := (scanArgs[i]).(*sql.NullBool); ok {
-						masterData[v.Name()] = z.Bool
-						continue
-					}
-
-					if z, ok := (scanArgs[i]).(*sql.NullString); ok {
-						masterData[v.Name()] = z.String
-						continue
-					}
-
-					if z, ok := (scanArgs[i]).(*sql.NullInt64); ok {
-						masterData[v.Name()] = z.Int64
-						continue
-					}
-
-					if z, ok := (scanArgs[i]).(*sql.NullFloat64); ok {
-						masterData[v.Name()] = z.Float64
-						continue
-					}
-
-					if z, ok := (scanArgs[i]).(*sql.NullInt32); ok {
-						masterData[v.Name()] = z.Int32
-						continue
-					}
-
-					masterData[v.Name()] = scanArgs[i]
-				}
-
-				sqlData.Data = append(sqlData.Data, masterData)
-				rowCount += 1
-			}
-
-			logger.LogMsg(fmt.Sprintf("found : %d records", rowCount), "info")
-			sqlData.ReturnedRows = rowCount
-
-			checkErr(rows.Close())
-			checkErr(json.NewEncoder(w).Encode(sqlData))
-			checkErr(db.Close())
-
-		} else {
-
-			_, err := io.Copy(w, endResponse)
-			checkErr(err)
-
-		}
-
-	} else {
-
-		endResponse := strings.NewReader(
-			fmt.Sprintf("{\"error\" : \"Sorry the call %s was undefined\"}\n", apiName),
-		)
-		_, err := io.Copy(w, endResponse)
-		checkErr(err)
-		logger.LogMsg(fmt.Sprintf("Sorry the call %s was undefined", apiName), "info")
-
-	}
-
-}
-
-func createApiApi(apiName string) {
-	logger.LogMsg(fmt.Sprintf("Requested api endpoint : %s", apiName), "info")
-
-	apiMethod := viper.GetString(fmt.Sprintf("api.%s.method", apiName))
-
-	if apiMethod == "get" {
-		getR.HandleFunc(apiName, queryApi)
-		logger.LogMsg(fmt.Sprintf("Created GET api endpoint : %s", apiName), "info")
-	}
-
-}
-
-func queryApi(w http.ResponseWriter, r *http.Request) {
-
-	w.Header().Set("Content-Type", "application/json")
-	apiName := r.URL
-	logger.LogMsg(fmt.Sprintf("Call %s", r.URL), "info")
-
-	if viper.IsSet(fmt.Sprintf("api.%s", apiName)) {
-
-		var response *http.Response
-		var err error
-
-		apiSource := viper.GetString(fmt.Sprintf("api.%s.source", apiName))
-		apiTargetMethod := viper.GetString(fmt.Sprintf("api.%s.target_method", apiName))
-		apiTargetUrl := viper.GetString(fmt.Sprintf("sources.%s.url", apiSource))
-
-		switch apiTargetMethod {
-		case "get":
-			response, err = http.Get(apiTargetUrl)
-		//case "post":
-		//	response, err = http.Post(apiTargetUrl)
-		default:
-			logger.LogMsg(
-				fmt.Sprintf("Sorry but the http call %s is not recognized", apiTargetMethod),
-				"critical",
-			)
-
-		}
-
-		checkErr(err)
-
-		_, err = io.Copy(w, response.Body)
-		checkErr(err)
-
-		// Add an CR at the end of the json stream
-		endResponse := strings.NewReader("\n")
-		_, err = io.Copy(w, endResponse)
-		checkErr(err)
-		httpClose(response.Body)
-
-	} else {
-
-		endResponse := strings.NewReader(
-			fmt.Sprintf("{\"error\" : \"Sorry the call %s was undefined\"}\n", apiName),
-		)
-		_, err := io.Copy(w, endResponse)
-		checkErr(err)
-		logger.LogMsg(fmt.Sprintf("Sorry the call %s was undefined", apiName), "info")
-
-	}
-
-}
-
-func checkErr(err error) {
-	if err != nil {
-		logger.LogMsg(fmt.Sprintf("An error occured %s", err), "critical")
-		panic(err)
-	}
-}
 
 func createEndpoints() {
 	listApiDescriptors := viper.GetStringMap("api")
@@ -305,9 +22,9 @@ func createEndpoints() {
 		logger.LogMsg(fmt.Sprintf("Connect type: %s", apiType), "info")
 		switch apiType {
 		case "sql":
-			createApiSql(apiDescriptor)
+			apisql.CreateApiSql(apiDescriptor)
 		case "api":
-			createApiApi(apiDescriptor)
+			apiapi.CreateApiApi(apiDescriptor)
 		default:
 			logger.LogMsg(fmt.Sprintf("Sorry type : %s is not implemented", apiType), "warning")
 		}
@@ -319,16 +36,7 @@ func HandleRequests() {
 
 	var err error
 
-	// create a serve mux
-	//sm = mux.NewRouter()
-	sm.StrictSlash(true)
-	//myRouter := mux.NewRouter().StrictSlash(true)
-
-	// register handlers
-	//postR := sm.Methods(http.MethodPost).Subrouter()
-	//getR = sm.Methods(http.MethodGet).Subrouter()
-	//putR := sm.Methods(http.MethodPut).Subrouter()
-	//deleteR := sm.Methods(http.MethodDelete).Subrouter()
+	globalvar.Sm.StrictSlash(true)
 
 	viper.OnConfigChange(func(e fsnotify.Event) {
 		logger.LogMsg(fmt.Sprintf("Config file changed: %s", e.Name), "info")
@@ -337,28 +45,9 @@ func HandleRequests() {
 	viper.WatchConfig()
 	createEndpoints()
 
-	// Define GET Call
-	//getR.HandleFunc("/users", users.AllUsers)
-
-	// Define POST Call
-	//postR.HandleFunc("/user/{Username}", users.CreateUser)
-
-	// Define DELETE Call
-	//deleteR.HandleFunc("/user/{username}", users.DeleteUser)
-
-	// Define PUT Call
-	//putR.HandleFunc("/user/{username}/{email}", updateUser)
-
-	//// used the PathPrefix as workaround for scenarios where all the
-	//// get requests must use the ValidateAccessToken middleware except
-	//// the /refresh-token request which has to use ValidateRefreshToken middleware
-	//refToken := sm.PathPrefix("/refresh-token").Subrouter()
-	//refToken.HandleFunc("", uh.RefreshToken)
-	//refToken.Use(uh.MiddlewareValidateRefreshToken)
-
 	listeningPort := viper.GetString("listening_port")
 
-	err = http.ListenAndServe(":"+listeningPort, sm)
-	checkErr(err)
+	err = http.ListenAndServe(":"+listeningPort, globalvar.Sm)
+	globalvar.CheckErr(err)
 
 }
